updated notes

+++ backend/controllers/notes.js
@@ -1,6 +1,5 @@
 const notesRouter = require('express').Router()
 const Note = require('../models/note')
-const User = require('../models/user')
 
 notesRouter.get('/', async (request, response) => {
   const notes = await Note.find({}).populate('user', { username: 1, name: 1 })
@@ -10,18 +9,22 @@ notesRouter.get('/', async (request, response) => {
 
 notesRouter.post('/', async (request, response) => {
   const body = request.body
-  const user = await User.findById(body.userId)
+  const user = request.user
 
-  const note = new Note({
-    content: body.content,
-    important: body.important || false,
-    user: user.id,
-  })
-  const savedNote = await note.save()
-  user.notes = user.notes.concat(savedNote._id)
-  await user.save()
+  if (user) {
+    const note = new Note({
+      content: body.content,
+      important: body.important || false,
+      user: user._id,
+    })
+    const savedNote = await note.save()
+    user.notes = user.notes.concat(savedNote._id)
+    await user.save()
 
-  response.status(201).json(savedNote)
+    response.status(201).json(savedNote)
+  } else {
+    response.status(401).json({ error: 'token not provided' })
+  }
 })
 
 notesRouter.get('/:id', async (request, response) => {
@@ -35,22 +38,39 @@ notesRouter.get('/:id', async (request, response) => {
 })
 
 notesRouter.delete('/:id', async (request, response) => {
-  await Note.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const note = await Note.findById(request.params.id)
+
+  if (note.user.toString() === user._id.toString()) {
+    await Note.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'token invalid' })
+  }
 })
 
 notesRouter.put('/:id', async (request, response) => {
   const body = request.body
+  const note = await Note.findById(request.params.id)
+  const user = request.user
 
-  const note = {
-    content: body.content,
-    important: body.important,
+  if (note.user.toString() === user._id.toString()) {
+    const updatedNote = {
+      content: body.content,
+      important: body.important,
+    }
+
+    const returnedNote = await Note.findByIdAndUpdate(
+      request.params.id,
+      updatedNote,
+      {
+        new: true,
+      }
+    )
+    response.json(returnedNote)
+  } else {
+    response.status(401).json({ error: 'token invalid' })
   }
-
-  const updatedNote = await Note.findByIdAndUpdate(request.params.id, note, {
-    new: true,
-  })
-  response.json(updatedNote)
 })
 
 module.exports = notesRouter
